Move tests back to test_design
