+++ tests/integration/test_body.py
@@ -337,67 +337,6 @@ def test_body_tracker_update_paths(modeler: Modeler):
         pyansys_geo.USE_TRACKER_TO_UPDATE_DESIGN = original_tracker
 
 
-def test_vertex_repr(modeler: Modeler):
-    """Test Vertex.__repr__ coverage."""
-    sketch = Sketch()
-    sketch.box(Point2D([0, 0]), 1, 1)
-    design = modeler.create_design("VertexReprTest")
-    body = design.extrude_sketch("box", sketch, 1)
-
-    if len(body.vertices) > 0:
-        vertex = body.vertices[0]
-        repr_str = repr(vertex)
-        assert "Vertex" in repr_str
-        assert vertex.id in repr_str
-        assert "Position" in repr_str
-        assert body.id in repr_str
-
-
-def test_assigning_and_getting_material(modeler: Modeler):
-    """Test the assignment and retrieval of materials from a design."""
-    sketch = Sketch()
-    sketch.circle(Point2D([10, 10], UNITS.mm), Quantity(10, UNITS.mm))
-
-    design_name = "ExtrudeProfile"
-    design = modeler.create_design(design_name)
-
-    density = Quantity(125, 1000 * UNITS.kg / (UNITS.m**3))
-    poisson_ratio = Quantity(0.33, UNITS.dimensionless)
-    tensile_strength = Quantity(45.0, UNITS.pascal)
-    material = Material(
-        "steel",
-        density,
-        [MaterialProperty(MaterialPropertyType.POISSON_RATIO, "myPoisson", poisson_ratio)],
-    )
-    material.add_property(MaterialPropertyType.TENSILE_STRENGTH, "myTensile", Quantity(45))
-    design.add_material(material)
-
-    body = design.extrude_sketch("JustACircle", sketch, Quantity(10, UNITS.mm))
-
-    body.material = material
-    mat_service = body.material
-
-    assert mat_service.name == "steel"
-    assert len(mat_service.properties) == 3
-    assert mat_service.properties[MaterialPropertyType.DENSITY].type == MaterialPropertyType.DENSITY
-    assert mat_service.properties[MaterialPropertyType.DENSITY].name == "Density"
-    assert mat_service.properties[MaterialPropertyType.DENSITY].quantity == density
-    assert (
-        mat_service.properties[MaterialPropertyType.POISSON_RATIO].type
-        == MaterialPropertyType.POISSON_RATIO
-    )
-    assert mat_service.properties[MaterialPropertyType.POISSON_RATIO].name == "myPoisson"
-    assert mat_service.properties[MaterialPropertyType.POISSON_RATIO].quantity == poisson_ratio
-    assert (
-        mat_service.properties[MaterialPropertyType.TENSILE_STRENGTH].type
-        == MaterialPropertyType.TENSILE_STRENGTH
-    )
-    assert mat_service.properties[MaterialPropertyType.TENSILE_STRENGTH].name == "myTensile"
-    assert (
-        mat_service.properties[MaterialPropertyType.TENSILE_STRENGTH].quantity == tensile_strength
-    )
-
-
 def test_get_empty_material(modeler: Modeler):
     """Check that the material service returns an empty material."""
     sketch = Sketch()
@@ -501,91 +440,6 @@ def test_create_surface_from_copy_faces(modeler: Modeler):
     circular_surface = design.copy_faces("circular", [face])
     assert circular_surface.is_surface
     assert circular_surface.faces[0].area.m == pytest.approx(78.53981633974483, rel=1e-6, abs=1e-8)
-
-
-def test_component_body(modeler: Modeler):
-    """Test the different ``Component`` and ``Body`` creation methods."""
-    design_name = "ComponentBody_Test"
-    design = modeler.create_design(design_name)
-    assert design.name == design_name
-    assert design.design_id is not None
-    assert design.id is not None
-    assert design.parent_component is None
-    assert len(design.components) == 0
-    assert len(design.bodies) == 0
-    assert len(design.materials) == 0
-    assert len(design.named_selections) == 0
-    assert len(design.coordinate_systems) == 0
-
-    sketch = Sketch()
-    sketch.polygon(Point2D([10, 10], UNITS.mm), Quantity(10, UNITS.mm), sides=5)
-
-    name_extruded_body = "ExtrudedPolygon"
-    distance_extruded_body = Quantity(50, UNITS.mm)
-    body = design.extrude_sketch(
-        name=name_extruded_body, sketch=sketch, distance=distance_extruded_body
-    )
-
-    assert body.name == name_extruded_body
-    assert body.id is not None
-    assert body.is_surface is False
-    assert len(body.faces) == 7
-    expected_vol = sketch.faces[0].area.m * distance_extruded_body.m * 1e-9
-    assert body.volume.m == pytest.approx(expected_vol)
-    assert len(design.components) == 0
-    assert len(design.bodies) == 1
-    assert len(body.edges) == 15
-    assert len(body.vertices) == 10
-
-    planar_component_name = "PlanarBody_Component"
-    planar_component = design.add_component(planar_component_name)
-    assert planar_component.id is not None
-    assert planar_component.name == planar_component_name
-
-    planar_sketch = Sketch()
-    planar_sketch.ellipse(
-        Point2D([50, 50], UNITS.mm), Quantity(30, UNITS.mm), Quantity(10, UNITS.mm)
-    )
-    planar_component_surface_name = "PlanarBody_Component_Surface"
-    planar_body = planar_component.create_surface(planar_component_surface_name, planar_sketch)
-
-    assert planar_body.name == planar_component_surface_name
-    assert planar_body.id is not None
-    assert planar_body.is_surface is True
-    assert len(planar_body.faces) == 1
-    assert planar_body.volume == 0.0
-    assert len(planar_component.components) == 0
-    assert len(planar_component.bodies) == 1
-    assert len(design.components) == 1
-    assert len(design.bodies) == 1
-    assert len(planar_body.edges) == 1
-
-    assert planar_component.parent_component.id == design.id
-
-    comp_str = repr(planar_component)
-    assert "ansys.geometry.core.designer.Component" in comp_str
-    assert "Exists               : True" in comp_str
-    assert "N Bodies             : 1" in comp_str
-    assert "N Components         : 0" in comp_str
-    assert "N Coordinate Systems : 0" in comp_str
-
-
-def test_design_part(modeler: Modeler):
-    """Test to validate the designer part id, name, and setter for components and bodies."""
-    body1 = MasterBody(id="body1", name="First Only Body", grpc_client=modeler.client)
-    body2 = MasterBody(id="body2", name="Second Body in Component", grpc_client=modeler.client)
-    bodies = [body1]
-    part = Part(id="IDPart", name="NamePart", components=[], bodies=bodies)
-    masterpart = MasterComponent(id="PartMaster", name="Part Master", part=part)
-    assert masterpart.id == "PartMaster"
-    assert masterpart.name == "Part Master"
-    assert masterpart.__repr__()[0:50] == "MasterComponent(id=PartMaster, name=Part Master, t"
-    assert part.id == "IDPart"
-    assert part.name == "NamePart"
-    part.components = [body2]
-    assert part.components[0].name == "Second Body in Component"
-    part.bodies = body1
-    assert part.bodies.name == "First Only Body"
 
 
 def test_body_get_named_selections(modeler: Modeler):
@@ -725,220 +579,6 @@ def test_update_body_sets_properties(modeler: Modeler):
         design._update_body(body, body_info)
 
         debug_spy.assert_called()
-
-
-def test_delete_body_component(modeler: Modeler):
-    """Test for verifying the deletion of ``Component`` and ``Body`` objects.
-
-    Notes
-    -----
-    Requires storing scdocx file and checking manually (for now).
-    """
-    # Create your design on the server side
-    design = modeler.create_design("Deletion_Test")
-
-    # Create a Sketch object and draw a circle (all client side)
-    sketch = Sketch()
-    sketch.circle(Point2D([-30, -30], UNITS.mm), Quantity(10, UNITS.mm))
-    distance = Quantity(30, UNITS.mm)
-
-    #  The following component hierarchy is made
-    #
-    #           |---> comp_1 ---|---> nested_1_comp_1 ---> nested_1_nested_1_comp_1
-    #           |               |
-    #           |               |---> nested_2_comp_1
-    #           |
-    # DESIGN ---|---> comp_2 -------> nested_1_comp_2
-    #           |
-    #           |
-    #           |---> comp_3
-    #
-    #
-    # Now, only "comp_3", "nested_2_comp_1" and "nested_1_nested_1_comp_1"
-    # will have a body associated...
-    #
-    #
-
-    # Create the components
-    comp_1 = design.add_component("Component_1")
-    comp_2 = design.add_component("Component_2")
-    comp_3 = design.add_component("Component_3")
-    nested_1_comp_1 = comp_1.add_component("Nested_1_Component_1")
-    nested_1_nested_1_comp_1 = nested_1_comp_1.add_component("Nested_1_Nested_1_Component_1")
-    nested_2_comp_1 = comp_1.add_component("Nested_2_Component_1")
-    _ = comp_2.add_component("Nested_1_Component_2")
-
-    # Create the bodies
-    body_1 = comp_3.extrude_sketch(name="comp_3_circle", sketch=sketch, distance=distance)
-    body_2 = nested_2_comp_1.extrude_sketch(
-        name="nested_2_comp_1_circle", sketch=sketch, distance=distance
-    )
-    _ = nested_1_nested_1_comp_1.extrude_sketch(
-        name="nested_1_nested_1_comp_1_circle", sketch=sketch, distance=distance
-    )
-
-    # Let's start by doing something impossible - trying to delete body_1 from comp_1
-    comp_1.delete_body(body_1)
-
-    # Check that all the underlying objects are still alive
-    assert comp_1.is_alive
-    assert comp_1.components[0].is_alive
-    assert comp_1.components[0].components[0].is_alive
-    assert comp_1.components[0].components[0].bodies[0].is_alive
-    assert comp_1.components[1].is_alive
-    assert comp_1.components[1].bodies[0].is_alive
-    assert comp_2.is_alive
-    assert comp_2.components[0].is_alive
-    assert comp_3.is_alive
-    assert comp_3.bodies[0].is_alive
-
-    # Do the same checks but calling them from the design object
-    assert design.is_alive
-    assert design.components[0].is_alive
-    assert design.components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].bodies[0].is_alive
-    assert design.components[0].components[1].is_alive
-    assert design.components[0].components[1].bodies[0].is_alive
-    assert design.components[1].is_alive
-    assert design.components[1].components[0].is_alive
-    assert design.components[2].is_alive
-    assert design.components[2].bodies[0].is_alive
-
-    # Let's do another impossible thing - trying to delete comp_3 from comp_1
-    comp_1.delete_component(comp_3)
-
-    # Check that all the underlying objects are still alive
-    assert comp_1.is_alive
-    assert comp_1.components[0].is_alive
-    assert comp_1.components[0].components[0].is_alive
-    assert comp_1.components[0].components[0].bodies[0].is_alive
-    assert comp_1.components[1].is_alive
-    assert comp_1.components[1].bodies[0].is_alive
-    assert comp_2.is_alive
-    assert comp_2.components[0].is_alive
-    assert comp_3.is_alive
-    assert comp_3.bodies[0].is_alive
-
-    # Do the same checks but calling them from the design object
-    assert design.is_alive
-    assert design.components[0].is_alive
-    assert design.components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].bodies[0].is_alive
-    assert design.components[0].components[1].is_alive
-    assert design.components[0].components[1].bodies[0].is_alive
-    assert design.components[1].is_alive
-    assert design.components[1].components[0].is_alive
-    assert design.components[2].is_alive
-    assert design.components[2].bodies[0].is_alive
-
-    # Let's delete now the entire comp_2 component
-    comp_2.delete_component(comp_2)
-
-    # Check that all the underlying objects are still alive except for comp_2
-    assert comp_1.is_alive
-    assert comp_1.components[0].is_alive
-    assert comp_1.components[0].components[0].is_alive
-    assert comp_1.components[0].components[0].bodies[0].is_alive
-    assert comp_1.components[1].is_alive
-    assert comp_1.components[1].bodies[0].is_alive
-    assert not comp_2.is_alive
-    assert comp_3.is_alive
-    assert comp_3.bodies[0].is_alive
-
-    # Do the same checks but calling them from the design object
-    assert design.is_alive
-    assert len(design.components) == 2
-    assert design.components[0].is_alive
-    assert design.components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].bodies[0].is_alive
-    assert design.components[0].components[1].is_alive
-    assert design.components[0].components[1].bodies[0].is_alive
-    assert design.components[1].is_alive
-    assert design.components[1].bodies[0].is_alive
-
-    # Let's delete now the body_2 object
-    design.delete_body(body_2)
-
-    # Check that all the underlying objects are still alive except for comp_2 and body_2
-    assert comp_1.is_alive
-    assert comp_1.components[0].is_alive
-    assert comp_1.components[0].components[0].is_alive
-    assert comp_1.components[0].components[0].bodies[0].is_alive
-    assert comp_1.components[1].is_alive
-    assert not body_2.is_alive
-    assert not comp_2.is_alive
-    assert comp_3.is_alive
-
-    # Do the same checks but calling them from the design object
-    assert design.is_alive
-    assert design.components[0].is_alive
-    assert design.components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].is_alive
-    assert design.components[0].components[0].components[0].bodies[0].is_alive
-    assert design.components[0].components[1].is_alive
-    assert design.components[1].is_alive
-    assert design.components[1].bodies[0].is_alive
-
-    # Finally, let's delete the most complex one - comp_1
-    design.delete_component(comp_1)
-
-    # Check that all the underlying objects are still alive except for comp_2, body_2 and comp_1
-    assert not comp_1.is_alive
-    assert not comp_2.is_alive
-    assert comp_3.is_alive
-    assert comp_3.bodies[0].is_alive
-
-    # Do the same checks but calling them from the design object
-    assert design.is_alive
-    assert len(design.components) == 1
-    assert design.components[0].is_alive
-    assert design.components[0].bodies[0].is_alive
-
-    # Finally, let's delete the entire design
-    design.delete_component(comp_3)
-
-    # Check everything is dead
-    assert design.is_alive
-    assert len(design.components) == 0
-
-    # Try deleting the Design object itself - this is forbidden
-    with pytest.raises(ValueError, match="The design itself cannot be deleted."):
-        design.delete_component(design)
-
-    # Let's try out the representation methods
-    design_str = str(design)
-    assert "ansys.geometry.core.designer.Design" in design_str
-    assert "Name                 : Deletion_Test" in design_str
-    assert "N Bodies             : 0" in design_str
-    assert "N Components         : 0" in design_str
-    assert "N Coordinate Systems : 0" in design_str
-    assert "N Named Selections   : 0" in design_str
-    assert "N Materials          : 0" in design_str
-    assert "N Beam Profiles      : 0" in design_str
-    assert "N Design Points      : 0" in design_str
-
-    comp_1_str = str(comp_1)
-    assert "ansys.geometry.core.designer.Component" in comp_1_str
-    assert "Name                 : Component_1" in comp_1_str
-    assert "Exists               : False" in comp_1_str
-    assert "Parent component     : Deletion_Test" in comp_1_str
-    assert "N Bodies             : 0" in comp_1_str
-    assert "N Beams              : 0" in comp_1_str
-    assert "N Components         : 0" in comp_1_str
-    assert "N Design Points      : 0" in comp_1_str
-    assert "N Coordinate Systems : 0" in comp_1_str
-
-    body_1_str = str(body_1)
-    assert "ansys.geometry.core.designer.Body" in body_1_str
-    assert "Name                 : comp_3_circle" in body_1_str
-    assert "Exists               : False" in body_1_str
-    assert "Surface body         : False" in body_1_str
-    assert "Parent component     : Component_3" in body_1_str
-    assert "Color                : None" in body_1_str
-
 
 def test_single_body_translation(modeler: Modeler):
     """Test for verifying the correct translation of a ``Body``.
@@ -1744,7 +1384,6 @@ def test_shell_body(modeler: Modeler):
     assert success
     assert base.volume.m == pytest.approx(Quantity(0.488, UNITS.m**3).m, rel=1e-6, abs=1e-8)
     assert len(base.faces) == 12
-
 
 def test_named_selection_build_faces_from_metadata_returns_none_when_body_missing():
     """Test face metadata build returns None when a referenced body is unavailable."""
